rptest: add `get_cloud_storage_type_and_url_style()`
When parameterizing ducktape tests, it makes sense to join the cloud
storage type with the url style to ensure compatability- if the two
options were to be parameterized separately in a @matrix decorator,
we might run, for example, an ABS test with path style addressing.

`get_cloud_storage_type_and_url_style()` will generate the proper
list of compatible cloud storage types and url styles.