`config`: add `tombstone_retention_ms` to cluster properties
Adding support for `tombstone_retention_ms`, which is a cluster property
that is equivalent to Kafka's `log.cleaner.delete.retention.ms`, but is more precisely
named, and represents the retention time for tombstone records in compacted topics
in `redpanda`.

This controls the default value for the topic level property `delete.retention.ms`.