`storage`: adjust `may_have_compactible_records()` condition
Consider the case in which we have 3 segments:

S: [S1] [S2] [ S3 ]
K: |K1| |K2| | K1 |
V: |V1| |V2| |null|

The current condition for `num_compactible_records > 1` in
`may_have_compactible_records()` would result in these segments being removed
from the range used for window compaction, and prevent the tombstone value
for `K1` in `S3` from being applied to `K1` in `S1`.

This condition is mostly due to historical reasons, in which we didn't
want to have completely empty segments post compaction. This issue is solved
by the placeholder feature.

Adjust it to `num_compactible_records > 0` to allow the above case to work
as expected. This change should not have any other dramatic effects on the
process of compaction.

Also modify tests that use `may_have_compactible_records()` to reflect
the updated behavior.