`storage`: add `mark_segment_as_finished_window_compaction()`
We use `seg->mark_as_finished_window_compaction()` to indicate
that a segment has been through a full round of window compaction,
whether it is completely de-duplicated ("clean") or only partially
indexed (still "dirty").

Add `mark_segment_as_finished_window_compaction()` to `segment_utils`
as a helper function to help mark a segment as completed window compaction,
and whether it is "clean" (in which case we mark the `clean_compact_timestamp`
in the `segment_index`).