bazel: introduce gazelle
Add support for generating build files for transform-sdk and kreq-gen

We must patch gazelle because it doesn't support a go.work file not in
the root pointing to other relative files. Putting go.work in the root
of the redpanda repo breaks existing workflows and tooling. This is nice
in that it mirrors our rust deps as well.