bazel/gazelle: use go.work hack
This file is used to compute labels, which the processing they use don't
use real filesystem traversals, and we don't want this file to be valid
anyways because it breaks existing tooling. So use paths that don't
exist just as a way of mapping all the go modules we support in Bazel
for gazelle.