rptest: add iam roles error to allow list
In ClusterConfigTest we chose to use fake AWS credentials to test cloud
configuration multi-property validation. When running on non-AWS clouds
with metadata service running but not expecting AWS valid metadata API
calls, the API calls we make to get credentials will fail with
surprising errors. Add these errors to the allow list.

In this test we are only interested in the config validation behavior.

CORE-6965