`kafka`: add `override_if_not_default` overload for `tristate`
Similar to the `override_if_not_default` overload for `std::optional<>`.
This ensures that certain `tristate` properties may still display
`DEFAULT_CONFIG` instead of `DYNAMIC_TOPIC_CONFIG` in certain cases.

There are more cases to check with `tristate`, however, the cases in which
the override is considered equivalent to the default are:

1. Empty (`std::nullopt`) and the optional default is also empty (`std::nullopt`),
2. Disabled (`disable_tristate`) and the optional default is empty (`std::nullopt`).
3. Holds a value and the optional default holds the same value.