`storage`: check for `may_have_tombstones` in `self_compact_segment`
Segments that may have tombstone records in them should still be
considered eligible for self compaction.

An early return statement was missing a check for this condition.
Add it so that tombstones will be properly removed for a segment
eligible for removal.

Also adjusts a `vassert()` to account for this case.

(cherry picked from commit 1f5f3633ee14522c0962daf18500a972228c2f60)