bazel: add target for fuzzing testing
A new has_sanitizers build flag is added.
This flag needs to be set to show presence of sanitizers.

Added extra global linking argument "-stdlib=libc++". Without this,
"libfuzzer.a", which is included by adding "-fsanitizer=fuzzer",
introduces a dependency on "stdlibc++.so". This ends up with both
"libc++.a" and "stdlibc++.so" being linked.

Added a config target to include debug symbols.

Added a test target for fuzz testing. Fuzz tests are only available
when the has_sanitizers flag is set.