omb: Limit cloud read throughput in TS read test
With the switch to m7gd we are now maxing out disk throughput when
catching up from the backlog.

This makes produce latency go bad. To avoid this limit cloud
readthroughput (as we do in cloud).

We use the same formula as in the cloud: 150MB/s / 12 * 2 (throughput /
total_shard_count * 2).