`features`: add `cloud_storage_backend_inference_removal` flag
Using inference for the cloud storage backend is deprecated as of `v25.1`.
Existing clusters which historically rely on inference may continue to use
it, but new clusters spun up must now specify `cloud_storage_backend` at the
cluster level.

This is an "anti"-feature flag, in that it indicates that post `v25.1`, this
code path is no longer permitted.