`storage`: add `_housekeeping_lock` to `disk_log_impl`
This lock will control concurrency between `gc()` and `housekeeping()`
in future commits. These are two functions that in the past have not
been concurrent.

We are locking to invoke these concurrently in two separate fibres
from the `log_manager`. It is expected that `gc()` is a fast process,
while `housekeeping()` (which performs compaction), is not.