`storage`: add `log::try_gc()`
This function attempts to obtain the `disk_log_impl::_housekeeping_lock`
right away in order to schedule urgent garbage collection.

If `try_get_units()` does not return any units, it means that there
is a concurrently running housekeeping fiber for the current log.

In this case, we early return so as not to block the `gc_loop()` in
`log_manager` from making progress. We will garbage collect on the next
loop of `housekeeping_loop()`, or try to garbage collect on the next
space management alert.