`storage`: add `compaction_lag_timestamp()`
A helper function which returns the timestamp of the first record
that has not been made cleanly compacted by sliding window compaction.

This along with `max.compaction.lag.ms` defines the time horizon for
forcing compaction of a log, in the event that the `dirty_ratio()`
has not yet reached `min.cleanable.dirty.ratio` in the defined
time bound.