`storage`: add `_compaction_lag_offset`
This is the offset that defines, inclusively, the last record that
was made cleanly compacted by sliding window compaction. That means
that the next offset is the first "dirty" (in the compaction sense)
offset.

We assign this value when rounds of sliding window compaction are
driven to completion. In order to do this, another offset,
`_last_compaction_window_end_offset` is added to book-keep the "start"
of the latest sliding window compaction round. Not to be confused
with `_last_compaction_window_start_offset`, which book-keeps how far
along the current round of sliding window compaction is.

Here is an example which could better explain what this means:

With notation;

D: A dirty segment
C: A cleanly compacted segment
S: `_last_compaction_window_start_offset`
E: `_last_compaction_window_end_offset`
L: `_compaction_lag_offset`

Starting with 5 dirty segments for the first round of sliding window
compaction, we have:

     [D] [D] [D] [D] [D]
  L--^                 ^--E

If we only fully indexed 2 dirty segments, we would have:

     [D] [D] [D] [C] [C]
  L--^        S--^     ^--E

If we drive sliding window compaction to completion, before making any
final updates to our offsets, we have:

     [C] [C] [C] [C] [C]
S/L--^                 ^--E

The `_lag` offset then takes the value of `_end`, and the sliding window
offsets are reset for the next round of window compaction.

     [C] [C] [C] [C] [C]
                       ^--L

If more segments were added, but the `dirty_ratio` of the log was not
high enough to force compaction, the log would look like this:

     [C] [C] [C] [C] [C] [D] [D] [D]
                       ^--L

Here, the `_lag` offset helps to define the time horizon for forcing
compaction of the first [D] segment after `max.compaction.lag.ms`.