`storage`: consider `max.compaction.lag.ms` in `apply_segment_ms()`
Per KIP-354, the force roll of compacted segments depends on the minimum
of `log.segment.ms` and `max.compaction.lag.ms` in order to provide
the proper guarantee for `max.compaction.lag.ms`.

Take the minimum between these two configurations in `apply_segment_ms()`
when deciding whether to roll a segment for a compacted topic
or not.