`config`: add `replica_lag_time_max_ms`
The maximum amount of time a follower may lag behind a leader before
it is reset.

This is important for dealing with potential race conditions between
compaction & recovery- to that effect, `replica_lag_time_max_ms`
should ALWAYS be less than `delete.retention.ms` for a topic to guarantee
safety.