`storage`: add `has_transaction_batches` to `index_state`
Answers the question of "does this segment have any transactional batches
in it?"

Just like `may_have_tombstone_records`, we need this in order to determine
whether or not a `segment` should undergo compaction in a timely manner
in order to remove transaction batches within. Just like the presence of
removable tombstones, the presence of removable transactional batches will act
as a forcing function for self compaction to occur (even if the `segment` is
already self compacted).

This is a definite flag, because we can quickly query whether a batch
contains transactional data or not in the append path (i.e we can default
this to `false` and only set it to `true` if a transactional batch is appended
to the `segment`).

However, this means that existing `segment`s on disk have a default value of
`false` for this flag. To be correct about its value (since `has` is a
definite prefix and `may_have` is un-definite), we iterate over the batches
via a `contains_tx_consumer` and set the flag if a transactional batch is found.

This ensures that `self_compact_timestamp` is only set if `has_transaction_batches`
has a value that is guaranteed to be correct.