`config`: rework `delete.retention.ms` validation
Because we will be using the configured value for `delete.retention.ms`
for the removal of transaction markers as well, we should no longer limit
the enablement of this topic/cluster property based on tiered storage
being disabled.

Safeguards are still in place in `ntp_config::tombstone_retention_ms()`
to ensure this value is not set if tiered storage properties are enabled.

Also update the cluster default value for `tombstone_retention_ms` to `24h`
(one day) to match Kafka's default.