`storage`: add `enhanced_key` utility functions
In order to get the `is_control` and `record_batch_type` fields back from
an `compacted_index::entry` which has been encoded via `enhance_key()` and
written to the `compacted_index` for a `segment`.

This will be used in a future commit to determine whether a record needs
to have its key appended to the `key_offset_map` built during sliding
window compaction.