`cluster`: add `shard_placement_table::remake_partition_state`
And add it to `current_state_marker` and `shard_local_state`.

This will indicate whether an `ntp` on the current shard needs to be
deleted and then recreated, and as such is persisted in the `kvstore`
via `current_state_marker` and in memory via the `shard_local_state`.

The plan is for the `ntp` reconciliation process in the `controller_backend`
to pick this flag up, and trigger the rebuild of an `ntp`.