`storage`: add `log::earliest_removable_timestamp()`
This function is going to be necessary for safety checks during recovery.

If a recovery round exceeds the time horizon set by
`earliest_removable_timestamp + delete.retention.ms`, there is the potential
for divergence across replicas due to removed state (i.e removed tombstone
records or transactional batches) not being seen by the follower.

The intention here is to, upon recovery start (at time of `recovery_stm`
construction) probe the `log` for its `earliest_removable_timestamp()` by
passing `follower->next_index`. This will return the earliest timestamp
for a record/batch above the offset from which recovery is being performed,
which will then be used in checks against the current time to recover in the
`recovery_stm`, resetting the learner state if necessary.