`storage`: add `log::max_removed_offset()`
This is another necessary piece of information to ensure divergence free
recovery.

If a round of recovery begins at an offset below `log->max_removed_offset()`,
there is the potential for divergence across replicas due to removed state
(i.e removed tombstone records or transactional batches) not being seen by
the follower.

The intention here is to perform one initial check at the start of recovery
(at time of `recovery_stm` construction) between the `follower->next_index`
and the `log->max_removed_offset()`. If `next < max_removed`, we must issue
a reset of the learner's state and begin recovery from the `log` start offset
once more.

The `max_removed_offset` is adjusted every time a tombstone or transactional
batch is removed due to the `delete.retention.ms` horizon being exceeded.
De-duplication (i.e removal due to a later record with the same key existing in
the log) does not affect its value.

Because this piece of information (unlike `earliest_removable_timestamp()`) is
not reconstructable from iteration over the `segment`s in a `log`, it must be
persisted in the `kvstore` per `partition`.