`ct/l1`: fix pause/resume handling of backgrounded `_work_fut` future
It was possible before for a pause/resume request to a worker to race
due to the use of `std::exchange` leaving `_work_fut` in a disengaged
state, but perhaps not a _resolved_ state.

Now we won't use `std::exchange`, but rather only clear `_work_fut` if
it has already been resolved. As well, in `resume_worker()`, if the `_work_fut`
still has a value, it is because a request to `pause_worker()` is still in
progress `co_await`'ing `_work_fut` before clearing it.

Add a conditional variable `_work_fut_cv` to allow the `pause_worker()` request
to alert the `resume_worker()` request once `_work_fut` has been resolved
and `reset()`, preventing any sort of race condition or multiple backgrounded
`_work_fut` fibers.